requirements.txt
aniso8601==9.0.1
attrs==22.2.0
autopep8==2.0.2
boto3==1.26.85
botocore==1.29.85
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
flask-restx==1.1.0
gunicorn==20.1.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
numpy==1.24.2
pandas==1.5.3
protobuf==3.20.3
pycodestyle==2.10.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==2.0.5.post1
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.15.0
